refactor: make model info enrichment provider-specific
