+++ README.md
@@ -17,7 +17,7 @@ Originally inspired by [opencode-lmstudio](https://github.com/nicktasios/opencod
 - **Auto-Injection**: Automatically adds unconfigured models into OpenCode provider config
 - **Provider Filtering**: Include or exclude specific providers from discovery
 - **Model Filtering**: Use regex rules to precisely control which discovered models are injected
-- **Model Metadata Enrichment**: Reads LiteLLM-style model metadata by default to populate context limits and reasoning variants
+- **Model Metadata Enrichment**: Optionally reads provider-specific model metadata to populate context limits and reasoning variants
 - **Non-Chat Filtering**: Skips models reported as non-chat by provider metadata, such as embeddings or image generation
 - **Configurable Discovery**: Control discovery behavior with global and provider-level enable/disable switches
 - **Smart Model Formatting**: Optional human-friendly display names for discovered models
@@ -104,7 +104,8 @@ Each provider can override discovery behavior through `provider.<name>.options.m
 |--------|------|-------------|
 | `provider.<name>.options.modelsDiscovery.enabled` | `boolean` | Override global discovery and provider filters for a single provider |
 | `provider.<name>.options.modelsDiscovery.endpoint` | `string` | Provider-specific models endpoint path. Defaults to `/v1/models` |
-| `provider.<name>.options.modelsDiscovery.modelInfoEndpoint` | `string` | Provider-specific model info endpoint path. Defaults to `/v1/model/info`; set to an empty string to disable metadata enrichment |
+| `provider.<name>.options.modelsDiscovery.modelInfoEndpoint` | `string` | Provider-specific model info endpoint path. Metadata enrichment is disabled when omitted |
+| `provider.<name>.options.modelsDiscovery.modelInfoFormat` | `string` | Model info response format used to parse and inject metadata. Currently supports `"litellm"` |
 | `provider.<name>.options.modelsDiscovery.filterNonChat` | `boolean` | When model info is available, skip models whose `model_info.mode` is not `chat`. Defaults to `true`; set to `false` to keep all discovered models |
 | `provider.<name>.options.modelsDiscovery.models.includeRegex` | `string[]` | Provider-specific model include filter |
 | `provider.<name>.options.modelsDiscovery.models.excludeRegex` | `string[]` | Provider-specific model exclude filter |
@@ -127,7 +128,7 @@ Priority rules:
 
 #### Model Metadata Enrichment
 
-LiteLLM exposes a richer `/v1/model/info` endpoint in addition to the OpenAI-compatible `/v1/models` endpoint. The plugin queries `/v1/model/info` by default and uses that metadata when the provider supports it. Providers with unusual endpoint paths can override `modelInfoEndpoint`; providers that do not support metadata can disable it with an empty string.
+LiteLLM exposes a richer `/v1/model/info` endpoint in addition to the OpenAI-compatible `/v1/models` endpoint. Because this endpoint is not part of the generic OpenAI-compatible API contract, metadata enrichment is opt-in. Configure both `modelInfoEndpoint` and `modelInfoFormat` to enable it for a provider.
 
 ```json
 {
@@ -140,7 +141,9 @@ LiteLLM exposes a richer `/v1/model/info` endpoint in addition to the OpenAI-com
         "baseURL": "http://127.0.0.1:4000/v1",
         "modelsDiscovery": {
           "enabled": true,
-          "endpoint": "/v1/models"
+          "endpoint": "/v1/models",
+          "modelInfoEndpoint": "/v1/model/info",
+          "modelInfoFormat": "litellm"
         }
       },
       "models": {}
@@ -167,6 +170,7 @@ For providers with custom metadata paths or non-standard behavior:
         "baseURL": "http://127.0.0.1:9000/v1",
         "modelsDiscovery": {
           "modelInfoEndpoint": "/custom/model-info",
+          "modelInfoFormat": "litellm",
           "filterNonChat": false
         }
       }
