+++ src/types/plugin-config.ts
@@ -13,10 +13,13 @@ export interface PluginConfig {
   smartModelName?: boolean
 }
 
+export type ModelInfoFormat = 'litellm' | (string & {})
+
 export interface ProviderDiscoveryConfig {
   enabled?: boolean
   endpoint?: string
   modelInfoEndpoint?: string
+  modelInfoFormat?: ModelInfoFormat
   filterNonChat?: boolean
   models?: {
     includeRegex?: string[]
