+++ src/utils/model-info/types.ts
@@ -1,0 +1,8 @@
+export interface ModelInfoEnricher {
+  shouldSkipModel(modelId: string): boolean
+  applyModelInfo(modelConfig: any, modelId: string): void
+}
+
+export interface ModelInfoEnricherOptions {
+  filterNonChat: boolean
+}
